requirements.txt
# GenAIOps Monitoring and Tracing Labs Dependencies
# Core Azure AI Services
azure-ai-projects>=1.0.0
azure-ai-inference>=1.0.0
azure-identity>=1.15.0

# Monitoring and Observability
azure-monitor-opentelemetry>=1.2.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0

# Environment Management
python-dotenv>=1.0.0
chromadb>=0.4.0

# Visualization for Labs
matplotlib>=3.6.0
plotly>=5.15.0

